boot: pxe_utils: Fix incorrect allocation of overlay_dir
The function label_boot_extension() calls snprintf on a buffer
overlay_dir. However, this buffer is not initialized properly. In the
"if" branch it is initialized using a variable "len", which is smaller
than the resulting buffer. In the "else" branch this variable is not
initialized at all. This results in a crash when calling a "sysboot"
command.

This commit fixes this issue.

Signed-off-by: Surkov Kirill <fanra3.tk@gmail.com>